+++ angular/lib/src/bootstrap/run.dart
@@ -185,7 +185,7 @@ ComponentRef<T> runApp<T>(
   }
   final injector = appInjector(createInjector);
   final ApplicationRef appRef = unsafeCast(injector.get(ApplicationRef));
-  return appRef.bootstrap(componentFactory, injector);
+  return appRef.bootstrap(componentFactory);
 }
 
 /// Asynchronous alternative to [runApp], supporting [beforeComponentCreated].
@@ -212,7 +212,7 @@ Future<ComponentRef<T>> runAppAsync<T>(
   final injector = appInjector(createInjector);
   return beforeComponentCreated(injector).then((_) {
     final appRef = unsafeCast<ApplicationRef>(injector.get(ApplicationRef));
-    return appRef.bootstrap(componentFactory, injector);
+    return appRef.bootstrap(componentFactory);
   });
 }